// src/types/intake.ts
export interface CaseIntake {
  id: number;
  clientName: string;
  dateOfLoss: string;
  caseType: string;
}

export interface IntakeData {
  injuries: string | null;
  id: string;
  clientName: string;
  gender: string | null;
  phoneNumber: string | null;
  email: string;
  address: string | null;
  city: string | null;
  zip: string | null;
  dateOfBirth: string | null;
  accidentDate: string | null;
  accidentTime: string | null;
  accidentLocation: string | null;
  accidentDescription: string | null;
  defendant1Name: string | null;
  defendant1Phone: string | null;
  defendant1Address: string | null;
  defendant1Carrier: string | null;
  defendant1CarrierPhone: string | null;
  defendant1Policy: string | null;
  defendant1Year: string | null;
  defendant1Make: string | null;
  defendant1Model: string | null;
  defendant1Damage: string | null;
  defendant2Name: string | null;
  defendant2Phone: string | null;
  defendant2Address: string | null;
  defendant2Carrier: string | null;
  defendant2CarrierPhone: string | null;
  defendant2Policy: string | null;
  defendant2Year: string | null;
  defendant2Make: string | null;
  defendant2Model: string | null;
  defendant2Damage: string | null;
  autoName: string | null;
  autoPhone: string | null;
  autoAddress: string | null;
  autoCarrier: string | null;
  autoAgent: string | null;
  autoPolicy: string | null;
  healthCarrier: string | null;
  healthPhone: string | null;
  healthType: string | null;
  healthAddress: string | null;
  healthGroup: string | null;
  healthPolicy: string | null;
  ambulance: boolean | null;
  ambulanceCompany: string | null;
  admitted: boolean | null;
  lengthOfStay: string | null;
  doctorHospital1: string | null;
  address1: string | null;
  phone1: string | null;
  treatmentDate1: string | null;
  doctorHospital2: string | null;
  address2: string | null;
  phone2: string | null;
  treatmentDate2: string | null;
  doctorHospital3: string | null;
  address3: string | null;
  phone3: string | null;
  treatmentDate3: string | null;
  bodyPartsAffected: string | null;
  priorInjuries: string | null;
  priorInsuranceClaims: string | null;
  priorAttorneys: string | null;
}